-로그인 사용자 정보 읽기 수정(시큐리티), (회원 API Controller)

+++ src/main/java/com/PetSitter/controller/Member/api/MemberApiController.java
@@ -1,5 +1,6 @@
 package com.PetSitter.controller.Member.api;
 
+import com.PetSitter.domain.Member.MemberDetails;
 import com.PetSitter.dto.Member.request.AddMemberRequest;
 import com.PetSitter.dto.Member.request.UpdateMemberRequest;
 import com.PetSitter.service.Member.MemberService;
@@ -9,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
@@ -53,7 +55,11 @@ public class MemberApiController {
 
     @Operation(description = "특정 회원 조회 API")
     @GetMapping("/{memberId}/myPage")
-    public ResponseEntity<?> findMember(@PathVariable("memberId") long id) {
+    public ResponseEntity<?> findMember(@PathVariable("memberId") long id, @AuthenticationPrincipal MemberDetails memberDetails) {
+        if (memberDetails.getMember().getId() != id) {
+            return ResponseEntity.badRequest()
+                    .build();
+        }
         Object member = memberService.findById(id);
 
         return ResponseEntity.ok()
@@ -68,7 +74,11 @@ public class MemberApiController {
 
     @Operation(description = "회원 탈퇴 API")
     @DeleteMapping("/{memberId}")
-    public ResponseEntity<Void> deleteMember(@PathVariable("memberId") long id) {
+    public ResponseEntity<Void> deleteMember(@PathVariable("memberId") long id, @AuthenticationPrincipal MemberDetails memberDetails) {
+        if (memberDetails.getMember().getId() != id) {
+            return ResponseEntity.badRequest()
+                    .build();
+        }
         memberService.delete(id);
 
         return ResponseEntity.ok()
@@ -78,7 +88,7 @@ public class MemberApiController {
     @Operation(description = "회원 정보 수정 API")
     @PutMapping("/{memberId}")
     public ResponseEntity<?> updateMember(@PathVariable("memberId") Long id, @RequestBody @Valid UpdateMemberRequest request,
-                                               BindingResult result) {
+                                               @AuthenticationPrincipal MemberDetails memberDetails, BindingResult result) {
         if (result.hasErrors()) {
             Map<String, String> errors = new HashMap<>();
 
@@ -90,6 +100,12 @@ public class MemberApiController {
             return ResponseEntity.badRequest()
                     .body(errors);
         }
+
+        if (memberDetails.getMember().getId() != id) {
+            return ResponseEntity.badRequest()
+                    .build();
+        }
+
         Object updateMember = memberService.update(id, request);
 
         return ResponseEntity.ok()
